거의 완성

+++ flask_server/team_bc/views/answer_view.py
@@ -1,3 +1,4 @@
+import flask_login
 from flask import Blueprint, jsonify
 from flask import request
 import json
@@ -13,7 +14,7 @@ bp = Blueprint('answer', __name__, url_prefix='/comment')
 
 
 ## 댓글 게재
-@bp.route("/<int:rid>/create", methods=["POST"])
+@bp.route("/create/<int:rid>", methods=["POST"])
 @login_required
 def add_comment(rid):
     """Add a comment and return the comment with ID"""
@@ -25,7 +26,7 @@ def add_comment(rid):
     information = Information.query.get_or_404(user_id)
     creator = information.name
     from datetime import datetime
-    answer = Answer(content=content, create_date=datetime.now(), creator=creator)
+    answer = Answer(content=content, create_date=datetime.now(), creator=creator, user_id=user_id)
     question.answer_set.append(answer)
     from team_bc import db
     db.session.commit()
@@ -37,7 +38,6 @@ def add_comment(rid):
 @login_required
 def show_comment(rid):
     question = Question.query.get_or_404(rid)
-    print(question.answer_set)
     result = []
     for ans in question.answer_set:
         result.append(ans.to_dict())
@@ -45,14 +45,17 @@ def show_comment(rid):
 
 
 ## 댓글 삭제
-@bp.route("/<int:rid>/delete", methods=["POST"])
+@bp.route("/delete/<int:rid>", methods=["POST"])
 @login_required
 def delete_comment(rid):
     dic_data = json.loads(request.data)
     answer = Answer.query.filter(Answer.question_id == rid).filter(Answer.id == dic_data["id"]).one()
-    print(answer)
-    from team_bc import db
-    db.session.delete(answer)
-    db.session.commit()
-    return jsonify()
-
+    if str(flask_login.current_user.id) == str(answer.user_id):
+        from team_bc import db
+        db.session.delete(answer)
+        db.session.commit()
+        response = jsonify()
+    else:
+        response = jsonify()
+        response.status_code = 401
+    return response
